[move-function] Implement SILFunction::isDefer().
Returns true if this SILFunction must be a defer.

NOTE: This may return false for defer statements that have been
deserialized without a DeclContext. This means that this is guaranteed to
be correct for SILFunctions in Raw SIL that were not deserialized as
canonical. Thus one can use it for diagnostics.