[ShrinkBorrowScope] Fixed assert.
Previously, it was asserted that hoisting over a terminator succeeded
(because it is previously checked that it can be done).  That was
erroneously done by putting the call to hoist within an invocation of
the assert macro.  The result was that in release builds, the call was
not made.  Here, that is fixed by just calling the hoist function and,
in debug builds, asserting its return.

rdar://86809882