Query Equatable/Comparable Conformances to Build Typed Derived ASTs
This resolves a case of a semantic ambiguity dredged up in rdar://86780149. The problem there is a library author exported a retroactive conformance of a type to Equatable. That type then gained an Equatable conformance in its originating module, and all clients downstream of them broke. In the Equatable synthesis code, it broke a very particular way where we'd assemble a UDRE that the expression checker wouldn't actually be able to resolve to a singular decl. Because this was implicit AST, it also didn't consider that error a strong enough condition to not make it to codegen.

We already made sure the Equatable/Comparable conformance for the members/payloads of the types entering synthesis were valid. Just use those to construct well-typed AST instead of relying on the expression checker.