[move-function] Implement Closure Argument Dataflow.
This dataflow is used to determine if a closure/defer use of a value as an
inout_aliasable value is consume inclusive-or liveness requiring use in a
caller. This information is used to determine if a closure invocation or defer
at the source level acts as a use after move in the source level caller of a
moved var or if the moved var is reinited in the function.

This is done by:

1. Classifying all uses of the inout_aliasable argument as either init, consume,
   use.

2. We then perform two separate dataflows that summarize the effect of the
   closure upon the specific address argument:

   a. First for each individual use, we walk upwards along predecessors to the
      beginning of the function stopping at consumes, inits, and other uses. If
      our dataflow eventually reaches the entrance block, then we know that we
      have a use that should be an error. If all uses are blocked by a consume,
      init, or other use then we know that if our operand was moved in a caller,
      the address is not used after the move at least in this callee.

   b. With that in hand, we then perform a similar dataflow for each individual
      consume operation except that we only stop at init blocks and other
      consume blocks. This leaves us with a set of consuming operations of the
      address that can be considered to be consume up out of the callee. We then
      make sure that these consumes form a post dominating set of upon the
      address or bail. If we bail, we do not handle the move in the caller
      causing us to emit an unhandled move error later after all diagnostics
      have run.

After performing both of these dataflows, our caller function can use our
outputs to emit an error if we had a use and a move is live in the caller or
create a new callee where our address argument is converted from an
inout_aliasable parameter to an out parameter.