[NFC] [ShrinkBorrowScope] Moved check.
Previously, the body--of the loop that iterated the instructions in the
block--contained the check that the instruction over which hoisting
might be done is the introducer.  Here, that check is moved into
tryHoistOverInstruction.  Now that function and its other caller,
canHoistOverInstruction, will return the right value (false) if they are
ever called with the introducer (though that could not happen without
this change).