Revert "SR-14635: Casts to NSCopying should not always succeed (#37683)"
This reverts commit c1fce93d0ec42f4440e0995e58cb90e4f9813182.

The stricter semantics for casting to a protocol-constrained class existential
seems to break some existing code.

So we're going to revert that in the 5.6 branch to eliminate problems.

Note:  The new semantics will remain in the main branch with the usual
backwards-compatibility allowances so that we can continue to improve
the correctness of the casting behavior.