utils: adjust the LLVM build
This adjusts the build invocation for the LLVM portion on Unix
platforms.  With the imminent rebranch, the new LLVM will pull
in a new version of google-benchmark which requires C++20 support
in the C++ compiler which rules out some of the currently enabled
targets.  Disable the benchmarks which are not used currently any
way.