Emit a separate vtable entry + override thunk for async overrides that drop throws
Unlike with synchronous functions, an async 'throws' function and an
async non-throwing function are not ABI-compatible. This would lead to
an ABI mismatch and runtime crash when overriding an `async throws`
function with an `async` one. Deal with this ABI incompatibility by
producing an appropriate thunk.

Fixes SR-15766 / rdar://problem/88035174.